backend/requirements.txt
langgraph
langchain
langchain-aws
boto3
fastapi
uvicorn
pydantic
python-dotenv
loguru
ddgs
wikipedia
beautifulsoup4
requests
fast-flights
sqlalchemy
asyncpg

